perf: ⚡️ Incluir atualizações
Inclusão de pacotes Snap e Flatpak na lista de atualização do sistema controlada pelo usuário.

+++ scripts/up_date_grade_clean.sh
@@ -1,10 +1,59 @@
 #!/usr/bin/env bash
 
 # Atuliazação do sistema
-sudo apt update
-sudo apt upgrade -y
-sudo apt full-upgrade -y
-sudo apt dist-upgrade -y
-sudo apt autoremove -y
-sudo apt autoclean -y
-sudo apt clean -y
+# Atualização de pacotes apt
+echo "Inicio da atualização do sistema"
+echo " "
+echo " "
+
+echo "Atualizar pacotes apt? (s/n)"
+read -r APT
+if [ "$APT" = "s" ]; then
+	sudo apt update
+	clear
+	sudo apt upgrade -y
+	clear
+	sudo apt full-upgrade -y
+	clear
+	sudo apt dist-upgrade -y
+	clear
+	sudo apt autoremove -y
+	clear
+	sudo apt autoclean -y
+	clear
+	sudo apt clean -y
+	echo " "
+	echo " "
+	echo "Pacotes apt atualizados com sucesso!"
+	echo " "
+	echo " "
+fi
+
+echo " "
+echo " "
+echo "Atualizar pacotes snap? (s/n)"
+read -r SNAP
+if [ "$SNAP" = "s" ]; then
+	sudo snap refresh
+	echo " "
+	echo " "
+	echo "Pacotes snap atualizados com sucesso!"
+	echo " "
+	echo " "
+fi
+echo " "
+echo " "
+echo "Atualizar pacotes flatpak? (s/n)"
+read -r FLATPAK
+if [ "$FLATPAK" = "s" ]; then
+	flatpak update
+	echo " "
+	echo " "
+	echo "Pacotes flatpak atualizados com sucesso!"
+	echo " "
+	echo " "
+fi
+echo " "
+echo " "
+clear
+echo "Sistema atualizado com sucesso!"
